modopenboundary: fix initialization
boundary(...)%uphase is used in openboundary_tend but
was set in subroutine openboundary_phasevelocity,
called in the time loop *after* openboundary_tend.

For now fixed by calling openboundary_phasevelocity near the end of
startup().  Note the placement is tricky, because it has to be after
rhobf has been initialized.

Also moved the allocation and initialization of rhointi from initopenboundary
to startup, so that it's done before the first openboundary_phasevelocity
but after rhobf.