modmicrophysics: stop and warn if Nc_0 is unreasonably small
The cloud droplet number concentration Nc_0 should be given
in number per m3. Giving it in the common unit of number
per cm3 results in a confusing and hard to track down error
from RRTMG at dawn (ask me how I know):

LIQUID SSA GRTR THAN 1.0

This commit adds a stop with a warning, if Nc_0 < 1e4 at
initialization,  which is a reasonably sure sign it was given
in the wrong unit.

There is probably a bug in RRTMG as well - it says it accepts
cloud droplet sizes up to 60 um, but a droplet > 59 um
turns what should be an interpolation in a table into
an extrapolation, and the extrapolated value exceeds the limit
1.0, which is checked for afterwards.