modsimpleice: fix uninitialized sgratio, loop up to kmax
fixes to bring modsimpleice closer to modsimpleice2.
Here fix uses of uninitialized variables.

* sgratio,rsgratio were calculated only for cells with qr.
In the sedimentation time splitting loop, those ratios were used also
for cells that initially did not have qr. The ratios are now calculated for
all cells.

* changed the process loops to go up to kmax, not k1.
which for modbulkmicro improved the stability with open boundary conditions.